- add adding track to playlist database, repo, interactor

+++ app/src/main/java/com/example/playlistmaker/library/data/db/PlaylistTrackEntity.kt
@@ -8,5 +8,5 @@ data class PlaylistTrackEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Int,
     val playlistId: Int,
-    val trackId: Int
+    val trackId: Long
 )
